moved the splash page picture into the new components folder
